server/bot/requirements.txt
python-telegram-bot
motor
python-dotenv
fastapi
httpx
uvicorn
pydantic

